+++ udata/app.py
@@ -207,12 +207,12 @@ def register_extensions(app):
         models, routing, tasks, mail, i18n, auth, theme, search, sitemap,
         sentry
     )
+    tasks.init_app(app)
     i18n.init_app(app)
     models.init_app(app)
     routing.init_app(app)
     auth.init_app(app)
     cache.init_app(app)
-    tasks.init_app(app)
     csrf.init_app(app)
     nav.init_app(app)
     theme.init_app(app)
@@ -220,7 +220,6 @@ def register_extensions(app):
     search.init_app(app)
     sitemap.init_app(app)
     sentry.init_app(app)
-    from . import patch_flask_security  # noqa
     return app