+++ udata/core/user/models.py
@@ -65,6 +65,9 @@ class User(WithMetrics, UserMixin, db.Document):
     # when SECURITY_CONFIRMABLE is True
     confirmed_at = db.DateTimeField()
 
+    password_rotation_demanded = db.DateTimeField()
+    password_rotation_performed = db.DateTimeField()
+
     # The 5 fields below are required for Flask-security
     # when SECURITY_TRACKABLE is True
     last_login_at = db.DateTimeField()