+++ udata/tests/dataset/test_dataset_tasks.py
@@ -3,8 +3,11 @@ import pytest
 from udata.models import Dataset, Topic, CommunityResource
 from udata.core.dataset import tasks
 from udata.core.dataset.factories import DatasetFactory, CommunityResourceFactory
-# csv.adapter for Tag won't be registered if this is not imported :thinking:
-from udata.core.tags import csv as _  # noqa
+# Those imports seem mandatory for the csv adapters to be registered. This might be because of the decorator mechanism.
+from udata.core.dataset.csv import DatasetCsvAdapter, ResourcesCsvAdapter, IssuesOrDiscussionCsvAdapter
+from udata.core.organization.csv import OrganizationCsvAdapter
+from udata.core.reuse.csv import ReuseCsvAdapter
+from udata.core.tags.csv import TagCsvAdapter
 
 
 pytestmark = pytest.mark.usefixtures('clean_db')
@@ -33,17 +36,6 @@ def test_purge_datasets_community():
     assert CommunityResource.objects.count() == 0
 
 
-def test_purge_orphan_community():
-    dataset = Dataset.objects.create(title='test_dataset')
-    community_resource1 = CommunityResourceFactory(title='test_community_1')
-    community_resource2 = CommunityResourceFactory(title='test_community_2')
-    community_resource1.dataset = dataset
-    community_resource1.save()
-
-    tasks.purge_orphan_community_resources()
-    assert CommunityResource.objects.count() == 1
-
-
 @pytest.mark.usefixtures('instance_path')
 def test_export_csv(app):
     dataset = DatasetFactory()
@@ -55,3 +47,5 @@ def test_export_csv(app):
     extras = [r.extras.get('csv-export:model') for r in dataset.resources]
     for model in models:
         assert model in extras
+    fs_filenames = [r.fs_filename for r in dataset.resources if r.url.endswith(r.fs_filename)]
+    assert len(fs_filenames) == len(dataset.resources)